Macro redefined::redefined_remote

redefined_remote!() { /* proc-macro */ }

REMOTE MACRO:

Macro Input

(Optional) #[derives(...)] for derives on new type

(Optional) other container attributes (i.e. #[...]) for on new type

Identifiers of the remote types (comma seperated surrounded by brackets: [A, B, ..])

the crate of the remote type as it appears in Cargo.toml

NOTE: If the type is part of a workspace, make sure the package referenced is the workspace package

Macro Output

A new type with the same fields as the old type

It’s name is the same expect with a concatenated Redefined

For nested types, you also need to call use the macro on each of the nested types

The new type automatically derives From<OLD TYPE> and Into<Old Type>

Note:

Currently only uses transmute to convert between the types, implement other methods

Example

    use ruint::Uint;

   redefined_remote!([Uint] : "ruint");
   redefined_remote!(#[derive(Clone)] [Uint] : "ruint");